restructure norm eq and add to ensemble
