+++ .gitignore
@@ -4,3 +4,8 @@ __pycache__/helpers.cpython-312.pyc
 __pycache__/helpers.cpython-313.pyc
 __pycache__/realgm_scraper.cpython-313.pyc
 __pycache__
+combined_player_data_with_labels_TEST_DATA.csv
+combined_player_data_with_labels.csv
+combined_player_data.csv
+PRE-TESTED.csv
+TESTED.csv
